@@ -5,25 +5,25 @@
   "private": true,
   "dependencies": {
     "@ant-design/icons": "^5.2.6",
-    "@types/jest": "^27.5.2",
+    "@types/jest": "^29.5.11",
     "@types/lodash": "^4.14.202",
     "@types/node": "^20.10.5",
-    "@types/react": "^17.0.62",
-    "@types/react-dom": "^17.0.17",
+    "@types/react": "^18.2.45",
+    "@types/react-dom": "^18.2.18",
     "@types/react-highlight-words": "^0.16.7",
     "@types/react-router-dom": "^5.3.3",
     "animate.css": "^4.1.1",
     "antd": "^5.12.5",
     "clsx": "^2.0.0",
     "firebase": "^10.7.1",
-    "konva": "^8.3.9",
+    "konva": "^9.3.0",
     "lodash": "^4.17.21",
     "moment": "^2.29.4",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-highlight-words": "^0.20.0",
     "react-hooks-global-state": "^2.1.0",
-    "react-konva": "^17.0.2-6",
+    "react-konva": "^18.2.10",
     "react-query": "^3.39.3",
     "react-router-dom": "6.21.1",
     "react-scripts": "5.0.1",
@@ -86,9 +86,9 @@
     ]
   },
   "devDependencies": {
-    "@testing-library/jest-dom": "^5.16.4",
-    "@testing-library/react": "^12.1.5",
-    "@testing-library/user-event": "^13.5.0",
+    "@testing-library/jest-dom": "^6.1.5",
+    "@testing-library/react": "^14.1.2",
+    "@testing-library/user-event": "^14.5.1",
     "gh-pages": "^5.0.0",
     "husky": "^8.0.0",
     "lint-staged": "^13.2.3",
